機能: 関数にJSDocコメントを追加し、型定義を整理
- `fetchUtil.ts`、`formatTweetQueryParams.ts`、`convertDateLabelToDate.ts`、`renderAnnotation.ts`、`toCompatibleBounds.ts`にJSDocコメントを追加しました。
- `MapAnnotationType.ts`で`MarkerAnnotationData`型を定義し、`data`プロパティを更新しました。
