+++ src/utils/type/map/MapAnnotationType.ts
@@ -1,3 +1,16 @@
+export type MarkerAnnotationData = {
+  id: number
+  category: string
+  location: { lat: number; lng: number }
+  area: string
+  link: string
+  publishedAt: Date
+  markerImgUrl: string
+}
+
+/**
+ * マップアノテーションの型
+ */
 export type MapAnnotationType = {
   id: number
   category: string // ニュースのカテゴリ
@@ -8,10 +21,7 @@ export type MapAnnotationType = {
   sourceName: string | null // ニュースの発信元
   sourceUrl: string | null // 発信元のURL
   clusteringIdentifier: string
-  data: {
-    area: string
-    link: string
-  }
+  data: MarkerAnnotationData
   markerImgUrl?: string // マーカーの画像
   publishedAt: Date // ニュースの公開日
   createdAt: Date // DBに登録された日時
